apps/api/requirements.txt
# FastAPI 核心
fastapi==0.110.0
uvicorn[standard]==0.27.1
uvloop==0.19.0
python-multipart==0.0.9

# 数据库
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.1

# Redis
redis[asyncio]==5.0.1

# MinIO
minio==7.2.4

# 数据验证
pydantic==2.6.1
pydantic-settings==2.2.1
email-validator==2.1.0

# XGBoost 和数据处理
xgboost==2.0.3
scikit-learn==1.4.0
numpy==1.26.4

# 安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 工具
python-dotenv==1.0.1
httpx==0.26.0
pandas==2.2.0
aiofiles==23.2.1

# 测试
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0